% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tract_hunter.R
\name{combine_asu_groups_internal}
\alias{combine_asu_groups_internal}
\title{Combine touching ASU groups}
\usage{
combine_asu_groups_internal(tract_data, nb)
}
\arguments{
\item{tract_data}{Tract data with asunum assignments}

\item{nb}{Neighbor list of contiguity}
}
\value{
Updated tract data with merged groups
}
\description{
Internal helper that merges adjacent ASU groups into one.
}
\keyword{internal}
