% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_articulation_points}
\alias{cpp_articulation_points}
\title{Find articulation points in the induced subgraph on selected nodes}
\usage{
cpp_articulation_points(nb_local, selected)
}
\arguments{
\item{nb_local}{R List of 0-based integer neighbor vectors}

\item{selected}{logical vector (TRUE = node is selected / active)}
}
\value{
0-based integer vector of cut vertices
}
\description{
Find articulation points in the induced subgraph on selected nodes
}
