% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_asu_original.r
\name{reshape_for_dashboard}
\alias{reshape_for_dashboard}
\title{Reshape algorithm output for the dashboard}
\usage{
reshape_for_dashboard(all_tr, asu_sf)
}
\arguments{
\item{all_tr}{sf object of all tracts}

\item{asu_sf}{sf object of ASU tracts}
}
\value{
A named list used by the dashboard
}
\description{
Reshape algorithm output for the dashboard
}
\keyword{internal}
