% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tract_hunter.R
\name{tract_hunter_combine_groups}
\alias{tract_hunter_combine_groups}
\title{Merge touching ASU groups and update state}
\usage{
tract_hunter_combine_groups(state)
}
\arguments{
\item{state}{Internal state list from \code{tract_hunter_seed}}
}
\value{
Updated state list
}
\description{
Merge touching ASU groups and update state
}
\keyword{internal}
