% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tract_hunter.R
\name{tract_hunter_finalize}
\alias{tract_hunter_finalize}
\title{Finalize Tract Hunter results}
\usage{
tract_hunter_finalize(state)
}
\arguments{
\item{state}{Internal state list from \code{tract_hunter_seed}}
}
\value{
List of objects for the dashboard
}
\description{
Finalize Tract Hunter results
}
\keyword{internal}
