% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tract_hunter.R
\name{tract_hunter_seed}
\alias{tract_hunter_seed}
\title{Initialize Tract Hunter state}
\usage{
tract_hunter_seed(
  tract_list,
  bls_df,
  ur_thresh = 0.0645,
  pop_thresh = 10000,
  verbose = TRUE,
  unemp_power = 0.9,
  ur_weight = 1
)
}
\arguments{
\item{tract_list}{List of tract data with contiguity information}

\item{bls_df}{BLS data frame}

\item{ur_thresh}{Unemployment rate threshold}

\item{pop_thresh}{Population threshold}

\item{verbose}{Logical, print progress messages}

\item{unemp_power}{Exponent applied to unemployment counts when scoring neighbors}

\item{ur_weight}{Power applied to unemployment rate when scoring neighbors}
}
\value{
A list representing the algorithm state
}
\description{
Sets up data structures required by the Tract Hunter algorithm.
}
\keyword{internal}
