% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASUbuildR-package.R
\docType{package}
\name{ASUbuildR-package}
\alias{ASUbuildR}
\alias{ASUbuildR-package}
\title{ASUbuildR: Build Areas of Substantial Unemployment}
\description{
ASUbuildR provides an interactive Shiny Flexdashboard plus supporting
algorithms for constructing Areas of Substantial Unemployment (ASUs) from
Census tract geometry and BLS workbook inputs.
}
\details{
The package includes multiple build strategies, including an optional
CP-SAT workflow backed by OR-Tools. For stable CP-SAT behavior, use
\code{setup_asu_python()} to provision the managed Python environment and
ensure OR-Tools \code{>= 9.15}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aisolori/ASUbuildR}
  \item Report bugs at \url{https://github.com/aisolori/ASUbuildR/issues}
}

}
\author{
\strong{Maintainer}: David Schmidt \email{deschmidt@detr.nv.gov}

Authors:
\itemize{
  \item Michael Frank
}

Other contributors:
\itemize{
  \item Antonio Solorio \email{aisolorio@detr.nv.gov} [contributor]
}

}
\keyword{internal}
