% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASUbuildR-package.R
\docType{package}
\name{ASUbuildR-package}
\alias{ASUbuildR}
\alias{ASUbuildR-package}
\title{ASUbuildR: A Shiny Interface for Constructing Areas of Substantial Unemployment}
\description{
Provides a point-and-click GUI together with helper algorithms for building Areas of Substantial Unemployment (ASUs) from Census tract and BLS data. The package wraps a Shiny Flexdashboard interface plus utilities for optimising tract groupings and exporting results.
}
\author{
\strong{Maintainer}: David Schmidt \email{deschmidt@detr.nv.gov}

Authors:
\itemize{
  \item Michael Frank
}

Other contributors:
\itemize{
  \item Antonio Solorio \email{aisolorio@detr.nv.gov} [contributor]
}

}
\keyword{internal}
