% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASUbuildR-package.R
\docType{package}
\name{ASUbuildR-package}
\alias{ASUbuildR}
\alias{ASUbuildR-package}
\title{ASUbuildR: Shiny interface to help state construct Areas of Substantial Unemployment}
\description{
This provides a GUI and related tools to help maximize unemployment within ASUs. Continuation lines should be indented.
}
\author{
\strong{Maintainer}: David Schmidt \email{deschmidt@detr.nv.gov}

Authors:
\itemize{
  \item Michael Frank
}

}
\keyword{internal}
