% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_asu_python.R
\name{check_asu_python}
\alias{check_asu_python}
\title{Check Python Setup for ASUbuildR}
\usage{
check_asu_python()
}
\value{
Logical. TRUE if properly configured, FALSE otherwise.
}
\description{
Checks if the Python environment is properly configured for ASUbuildR.
}
