\name{check_asu_python}
\alias{check_asu_python}
\title{Check Python Setup for ASUbuildR}
\usage{check_asu_python()}
\value{Logical TRUE if the environment is properly configured, otherwise FALSE.}
\description{Checks whether the ASUbuildR Python environment exists and has the
required packages installed.}
