% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tract_hunter.R
\name{run_tract_hunter}
\alias{run_tract_hunter}
\title{Execute the Tract Hunter algorithm}
\usage{
run_tract_hunter(
  tract_list,
  bls_df,
  ur_thresh = 0.0645,
  pop_thresh = 10000,
  join_touching = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{tract_list}{List of tract data with contiguity information}

\item{bls_df}{BLS data frame}

\item{ur_thresh}{Unemployment rate threshold}

\item{pop_thresh}{Population threshold}

\item{join_touching}{Logical, join touching ASUs after first pass}

\item{verbose}{Logical, print progress messages}
}
\value{
A list of results similar to \code{run_asu_original()}
}
\description{
This is a wrapper that runs all stages of the Tract Hunter algorithm
and returns the final results used by the dashboard.
}
\keyword{internal}
